+++ DT_MARS-EXPLORE/Mountainous_Mission.cpp
@@ -1,5 +1,5 @@
 #include "Mountainous_Mission.h"
-Mountainous_Mission::Mountainous_Mission(int TLOC_x, int MDUR_x, int SIG_x, int FD_x, int ID_x)
+Mountainous_Mission::Mountainous_Mission(double TLOC_x, int MDUR_x, int SIG_x, int FD_x, int ID_x)
 	:Mission(TLOC_x, MDUR_x, SIG_x, FD_x, ID_x)//?>=0 or > only?
 {
 	//Set_AutoP(AutoP_x);
